consolidate client analytics logic for mobile/desktop
