+++ src/desktop/analytics/timeOnPageListener.ts
@@ -1,3 +1,4 @@
+import { data as sd } from "sharify"
 import { trackEvent } from "./helpers"
 import { timeOnPage } from "@artsy/cohesion"
 import { getContextPageFromClient } from "lib/getContextPage"
@@ -34,4 +35,47 @@ export const timeOnPageListener = (delay: number = 15000) => {
       trackingOptions
     )
   }, delay)
+}
+
+class PageTimeTracker {
+  path: string
+  delay: number
+  description: string
+  timer: any
+
+  constructor(path, delay, description) {
+    this.path = path
+    this.delay = delay
+    this.description = description
+    this.timer = null
+    this.track()
+  }
+
+  setPath(newPath) {
+    this.path = newPath
+  }
+
+  track() {
+    this.timer = timeOnPageListener()
+  }
+
+  clear() {
+    if (this.timer) clearTimeout(this.timer)
+  }
+
+  reset(newPath = null) {
+    this.clear()
+    if (newPath) this.setPath(newPath)
+    this.track()
+  }
+}
+
+export const trackTimeOnPage = () => {
+  if (sd.IS_MOBILE) {
+    timeOnPageListener()
+  } else {
+    window.desktopPageTimeTrackers = [
+      new PageTimeTracker(sd.CURRENT_PATH, 15000, "15 seconds"),
+    ]
+  }
 }
